kernel/.../{krun,kprove,unparser}: new class KPrint which handles printing and writing files
