+++ kernel/src/main/java/org/kframework/unparser/ColorSetting.java
@@ -1,5 +1,5 @@
 // Copyright (c) 2014-2018 K Team. All Rights Reserved.
-package org.kframework.krun;
+package org.kframework.unparser;
 
 /**
  * @author Denis Bogdanas
